Fixes race condition
I ran into a race condition when trimming down the provided application.  If the frontend
bundle finished far enough ahead of the server, the server would fail to start (as that
bundle had not finished processing).

This PR waits for all configs to be bundled before moving onto the next step and solves
this race condition.

Minimal application that triggered the error is in my fork of the repo:
https://github.com/seriallos/react-starter-kit-minimal